+++ CHANGELOG.md
@@ -10,3 +10,10 @@
 ## 0.0.2
 
 - change config key `local-history` to `localHistory`
+
+## 0.0.3
+
+- cleanups
+- add new config `localHistory.suppressErrors`
+- update vscode minimum version to `1.68`
+- make sure to hide the tree items when no editor is opened